Adds Lattice connect component
This component is responsible for creating a connection to a user's Lattice
device using an instance of the GridPlus SDK. This instance is derived from
login info (device ID, password) which is saved to the store. The user can
rehydrate the SDK instance and automatically connect to the device's
active wallet whenever this component loads.
We do not store the SDK instance itself in the store - just information needed
to rehydrate it.
The user may switch wallets on the Lattice device using SafeCards. When this
happens the user can refresh the connection using this component and it will
change the context to that of the new active wallet.